fix: change variables name

+++ Challenge04/Challenge04/Model/CategoryTypes.swift
@@ -8,7 +8,7 @@
 import Foundation
 
 @objc
-public enum CategoryTypes: Int16 {
+public enum CategoryTypes: Int {
     case studies = 1
     case study = 2
     case work = 3
